Rename _pdg.py into _pdg_adapter.py to clarify its task
