+++ src/qrules/_pdg_adapter.py
